Move vcr code to common utils package
