+++ pkg/testutils/testutils.go
@@ -1,0 +1,17 @@
+package testutils
+
+import "strings"
+
+const (
+	BadSecret = "Bad_Secret"
+)
+
+func ErrorContains(out error, want string) bool {
+	if out == nil {
+		return want == ""
+	}
+	if want == "" {
+		return false
+	}
+	return strings.Contains(out.Error(), want)
+}
