+++ pkg/token/azurecli_test.go
@@ -1,14 +1,15 @@
 package token
 
 import (
-	"strings"
 	"testing"
+
+	"github.com/Azure/kubelogin/pkg/testutils"
 )
 
 func TestNewAzureCLITokenEmpty(t *testing.T) {
 	_, err := newAzureCLIToken("", "")
 
-	if !ErrorContains(err, "resourceID cannot be empty") {
+	if !testutils.ErrorContains(err, "resourceID cannot be empty") {
 		t.Errorf("unexpected error: %v", err)
 	}
 }
@@ -17,17 +18,7 @@ func TestNewAzureCLIToken(t *testing.T) {
 	azcli := AzureCLIToken{}
 	_, err := azcli.Token()
 
-	if !ErrorContains(err, "expected an empty error but received:") {
+	if !testutils.ErrorContains(err, "expected an empty error but received:") {
 		t.Errorf("unexpected error: %v", err)
 	}
-}
-
-func ErrorContains(out error, want string) bool {
-	if out == nil {
-		return want == ""
-	}
-	if want == "" {
-		return false
-	}
-	return strings.Contains(out.Error(), want)
 }
