+++ pkg/token/serviceprincipaltokencertificate_test.go
@@ -2,6 +2,8 @@ package token
 
 import (
 	"testing"
+
+	"github.com/Azure/kubelogin/pkg/testutils"
 )
 
 func TestMissingCertFile(t *testing.T) {
@@ -11,7 +13,7 @@ func TestMissingCertFile(t *testing.T) {
 	expectedErr := "failed to read the certificate file"
 
 	_, err := p.Token()
-	if !ErrorContains(err, expectedErr) {
+	if !testutils.ErrorContains(err, expectedErr) {
 		t.Errorf("expected error %s, but got %s", expectedErr, err)
 	}
 }
@@ -19,12 +21,12 @@ func TestMissingCertFile(t *testing.T) {
 func TestBadCertPassword(t *testing.T) {
 	p := &servicePrincipalToken{
 		clientCert:         "testdata/testCert.pfx",
-		clientCertPassword: badSecret,
+		clientCertPassword: testutils.BadSecret,
 	}
 	expectedErr := "failed to decode pkcs12 certificate while creating spt: pkcs12: decryption password incorrect"
 
 	_, err := p.Token()
-	if !ErrorContains(err, expectedErr) {
+	if !testutils.ErrorContains(err, expectedErr) {
 		t.Errorf("expected error %s, but got %s", expectedErr, err)
 	}
 }
